Refactor module to use metatable
This makes method declaration easier.  Additionally this decouples the
initialization process from everything else.  This was becoming a
problem when I was trying to reuse code in the initialization for
methods, so this way will allow me to get around that without changing
too much.  (other than making a neater code structure)