Fix staggered isometric maps tile positions
So this code was used from Simple Tiled Implmentation for Love2Ds
version.  In their code they start counting rows and columns at 1,
instead of Berry which is 0.  Due to this, the layout for all staggered
maps was off slightly.  By checking for the remainder of 1 instead of 0,
this fixes this issue.